show submodule imports for packages.
This allows for the following pattern:

    mypkg/__init__.py:
        from mypkg.submodule1 import a
        from mypkg.submodule2 import b

To generate documentation for `mypkg` that includes `a` and `b`.
Additionally, it adds a flag you can add to your `__init__.py`,
`__pdoc_nosubmodules`. When specified, documentation will not be
generated for any submodules or subpackages of that package. This allows
you to shadow inner modules of a package and present the contents of an
`__init__.py` as the sole API surface of a package.

While this was theoretically possible previously by setting `__all__` to
be the contents of your `__init__.py`, it quickly becomes tedious to
keep `__all__` up-to-date when a package is large, leading to holes in
the documentation.